Moved check function to helpers
