--- djangocms_version_locking/check.py
@@ -1,6 +1,0 @@
-from .helpers import content_is_unlocked
-
-
-def placeholder_content_is_unlocked(placeholder, user):
-    content = placeholder.source
-    return content_is_unlocked(content, user)
